Pulled registry out of the docker package
