+++ dfs/docker/mocks/Docker.go
@@ -98,8 +98,17 @@ func (_m *Docker) PullImage(image string) error {
 	return r0
 }
 
-func (_m *Docker) TagImage(oldImage string, newImage string) {
-	_m.Called(oldImage, newImage)
+func (_m *Docker) TagImage(oldImage string, newImage string) error {
+	ret := _m.Called(oldImage, newImage)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(string, string) error); ok {
+		r0 = rf(oldImage, newImage)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
 }
 
 func (_m *Docker) RemoveImage(image string) error {
